Allocated a separate module, repository and viewmodel for the Schedule feature.
