+++ app/src/main/java/com/nailorsh/repeton/data/repositories/FakeRepetonRepository.kt
@@ -8,22 +8,9 @@ import com.nailorsh.repeton.model.Lesson
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.withContext
-import java.time.LocalDate
 import javax.inject.Inject
 
 class FakeRepetonRepository @Inject constructor() : RepetonRepository {
-    override suspend fun getLessons(): List<Lesson> = withContext(Dispatchers.IO) {
-//        delay(5000)
-        FakeLessonSource.loadLessons()
-    }
-
-    override suspend fun getLessons(day: LocalDate): List<Lesson> = withContext(Dispatchers.IO) {
-//        delay(2000)
-        FakeLessonSource.loadLessons().filter {
-            it.startTime.year == day.year && it.startTime.dayOfYear == day.dayOfYear
-        }
-    }
-
     override suspend fun getChats(): List<Chat> = withContext(Dispatchers.IO) {
         FakeChatsSource.generateChatList()
     }
