+++ app/src/main/java/com/nailorsh/repeton/domain/repositories/ScheduleRepository.kt
@@ -1,0 +1,9 @@
+package com.nailorsh.repeton.domain.repositories
+
+import com.nailorsh.repeton.model.Lesson
+import java.time.LocalDate
+
+interface ScheduleRepository {
+    suspend fun getLessons(): List<Lesson>
+    suspend fun getLessons(day: LocalDate): List<Lesson>
+}
