+++ app/src/main/java/com/nailorsh/repeton/domain/viewmodels/RepetonViewModel.kt
@@ -13,7 +13,6 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import java.io.IOException
 import java.net.HttpRetryException
-import java.time.LocalDate
 import javax.inject.Inject
 
 
@@ -21,12 +20,6 @@ sealed interface CurrentLessonUiState {
     data class Success(val lesson: Lesson) : CurrentLessonUiState
     object Error : CurrentLessonUiState
     object Loading : CurrentLessonUiState
-}
-
-sealed interface ScheduleUiState {
-    data class Success(val lessons: Map<LocalDate, List<Lesson>>) : ScheduleUiState
-    object Error : ScheduleUiState
-    object Loading : ScheduleUiState
 }
 
 sealed interface ChatsUiState {
@@ -39,54 +32,12 @@ sealed interface ChatsUiState {
 class RepetonViewModel @Inject constructor(
     private val repetonRepository: RepetonRepository
 ) : ViewModel() {
-    var scheduleUiState: ScheduleUiState by mutableStateOf(ScheduleUiState.Loading)
-        private set
 
     var chatsUiState: ChatsUiState by mutableStateOf(ChatsUiState.Loading)
         private set
 
     var currentLessonUiState: CurrentLessonUiState by mutableStateOf(CurrentLessonUiState.Loading)
         private set
-
-    private var lessonsCache = mutableMapOf<LocalDate, MutableList<Lesson>>()
-
-
-    init {
-        getLessons()
-    }
-
-    fun getLessons() {
-        viewModelScope.launch {
-            scheduleUiState = ScheduleUiState.Loading
-            try {
-                val lessons = repetonRepository.getLessons()
-                lessons.forEach { lesson ->
-                    val day = LocalDate.from(lesson.startTime)
-                    lessonsCache[day] = (lessonsCache[day] ?: mutableListOf()).also {
-                        it.add(lesson)
-                    }
-                }
-                val todayLessons = lessonsCache ?: emptyMap<LocalDate, MutableList<Lesson>>()
-                scheduleUiState = ScheduleUiState.Success(todayLessons)
-            } catch (e: Exception) {
-                scheduleUiState = ScheduleUiState.Error
-            }
-        }
-    }
-
-//    fun getLessons(day: LocalDate) {
-//        viewModelScope.launch {
-//            scheduleUiState = ScheduleUiState.Loading
-//            scheduleUiState = try {
-//                ScheduleUiState.Success(lessonsCache.getOrDefault(day, listOf()))
-//            } catch (e: IOException) {
-//                ScheduleUiState.Error
-//            } catch (e: HttpRetryException) {
-//                ScheduleUiState.Error
-//            }
-//
-//        }
-//    }
 
     fun getLesson(lessonId: Int) {
         viewModelScope.launch {
